Add CombatEntity*, get health bar working again
Big refactor preparing to get status effects displaying.

Renamed Entity to CombatEntity to be specific about the kind of entity
this interface is for. Then delineated between CombatEntityBaseStats
(implemented by Companion and Enemy, the scriptable-object stored
persistent entities) and CombatEntityInEncounterStats, the latter of
which will do all of the math for calculating what a combat entity's
stats should be after all of its status effects. And keep track of those
status effects for CombatEntityStatsDisplay (The only one of these that
extends monobehavior) to display and make available to other scripts.

CombatEntityStatsDisplay finds the CombatEntityInEncounterStats it
should be displaying by finding a CombatEntityInstance in its parent (or
parent hierarchy), which both EnemyInstance and CompanionInstance
implement. The only method currently required of CombatEntityInstance is
getCombatEntityInEncounterStats.