Add Coroutine Targetting, Targetting all Valid Targets
Refactored targetting to use a coroutine for fetching effect targets so
that we have more consistent state. Got cancelling targetting working
with the "Q" key.

Added "Beeg boom" a card that deals damage to all enemies equal to the
caster's strength. Required adding reference to the enemy manager in the
card caster.