Preparations for implementing discard effects
Realized that the cards are going to be targettableEntities too, so did
a bunch of reworking of the class hierarchies. Also created
InCombatDeck, separate from Deck (which is what companions have and add
cards to) because 1) Deck doesn't  care about what's in a draw pile and
what's been discarded/exhausted 2) InCombatDeck cares about these things
and will have a lot of functions on it that Deck shouldn't care about
(deal cards from, discard card from, exhaust card).

So yeah just a bunch of refactoring with accompanying debugging and
fixing of the enemy and companion prefabs to handle the whole
Entity->TargettableEntity->CombatEntityInstance->Enemy/CompanionInstance
hierarchy we have going on now.

oh also **IMPORTANT NOTE**: It makes sense to have the entity id be a
field at the level of Entity for a lot of reasons, but companions have a
persistent id between scenes (and so will cards). So right now when the
entity (just talkin about enemies and companions rn) at the bottom of
the chain hits its Start method, it MUST set its own Entity.id to the id it has in its BaseStats.id field, or we will have some real big problems trying to load stuff into scenes in the future.