refactor variableSOs and start map and room implementation
This commit does a lot, most of it being refactoring the variable
scriptable objects. This abstracts the variable and creates a new
abstract reference class, based on the talk from Ryan Hipple from
unite 2017. This lets any reference to a map, room, or encounter in
be pulled either from a scriptable object variable or a constant value
set in the editor.

Something to note that might become confusing pretty quickly is how
scriptable objects are used. I've been using them two different ways,
the first way being to store rooms and encounters that I want to reuse.
These are intended essentially to be constants that are read only. The
second way is to use the SOs as data storage between scenes, like
setting the "ActiveEncounter" variable then using that variable to setup
the encounter scene. The only problem with this, is that they use the
same assets and can quickly become confused with one another. And
because they don't reset when leaving play mode, you might accidentally
override values that you had intended to be static if you're not
careful like I did. This is something I want to work on and fool proof
eventually, but will require some thinking.