Refactor CardInfo into CardType and Card
Realized through a simple bug where the discard effect of one couldn't discard
a different card of the same type that we weren't actually using
cardId's the way I meant to when they were first put onto the cards.
CardInfo was serving as both the base template for the cards and the
data-tracking backend for the entities in the scene.

Created CardType (and remade all of the cardTypes in the editor, thanks
Unity), and made CardInfo into Card, the same kind of non-SO or
Monobehavior class that we've been passing around in situations like
Companion. Companions now keep track of their Cards, which all have
their own Id unique to their instance (rather than being unique to their card type) through their Deck, which has its List<Card>. StartingDeck is a collection of CardTypes, rather than a collection of Cards.

And the bug is fixed! Discard effects can't target the card that owns
the effect, but they can target other cards of the same type.