Solve ability concurrency, refactor targetting
Solved the problem of having abilities that trigger on specific turn
phases all happen at the same time by having the TurnManager register
triggers on each turn phase and actually invoke the triggers' attached
IEnumerables. Thinking we'll probably want to generalize this to work
for enemy attacks and maybe even other types of events we want to
trigger multiple things on.

Also realized we had an opportunity for cleaner code in the form of
TargetProvider and TargetRequester. CompanionAbility and
EffectProcedure were interacting with CompanionAbilityInvoker and
CardCaster - respectively - in very similar ways, leading to a lot of
duplicated code. Now CardCaster and CompanionAbilityInvoker are
TargetProviders, and CompanionAbility and EffectProcedure both extend
TargetRequester. Thinking this relationship will end up being used in
the future.