Enemy Controls Iteration 1
Added an EnemyBrain Scriptable Object that gets an intent, generated at
random from a list of behaviors on the scriptable object.

Modified EnemyInstance to provide context to the EnemyBrain, which will
in turn provide it to the behavior that's chosen on a given turn.

The idea behind the EnemyBrain is that we may want to have behaviors be
non-randomly selected for some enemies. Hexaghost in sts is a good
example, he goes through his behaviors sequentially.