EnemyBrain, Behaviors, Intent Declaration and Display
EnemyTypes now have an EnemyBrain scriptableObject associated with them. The EnemyBrain's job is to choose an EnemyBehavior from its list of EnemyBehaviors on each turn. Right now only random choice and sequentially going through the behaviors list is implented, but the idea is that in the future we could have more interesting logic for how an enemy chooses what it wants to do on a given turn.

An EnemyBehavior (examples would be "big attack" or "small
multi-attack" or "buff self" or "debuff ally") turns the enemy's stats
into an EnemyIntent that can then be displayed and acted upon.

Also changed the TargettingArrow so that it can display a little loop if
it's targetting itself, was really hard to tell that an enemy was
self-targetting for its intent.