Add CombatEntityWithDeck, Refactor Companion and Enemy Instantiation
Jumped through a lot of hoops so that minions and companions could share
their card handling code. PrefabInstantiator's static methods now call
each other up the entity hierarchy, and each sets up the baseStats field
on that level of the hierarchy. I think there could be a nice way to do
this recursively, but leaving that as a todo for now.