Setup for cardSelectionUI
Got the cards from a companion's deck showing up when they're targetted
with the Bellows effect. I don't think this is the last iteration here,
but I want something working before I put too much more thought into it.
Next step is to get the cards the CardViewUI displays to have a default
action onclick (exhaust, in this example) and when the view goes away
(put into hand).