First pass at reimplementing all card Effects with selection menu
Biggest problem we're running into here, and what's making the code the
most messy right now, is that the SelectionRequest even is actuated by
the CardSelectionManager. We should just pass the two lists back to
whoever requested the card targets and let them do what they will with
it. Started making the events (not filled in right now) that will do
that.