Work on abstracting all card effects to use TargetProvider
See TargetProvider.cs->requestCardTargets() and the coroutine it calls for most of the changes. Just trying to make it so that we follow the same target request/target supplied pattern for our card effects (things that target cards) that we use for combat effects (things that target combatEntities).

Dropped because the state (reset in resetTargettingState(), updated and checked on by muultiple coroutines in the class) in TargetProvider got really tangled up
between cardTargetRequests and effectTargetRequests. I believe it's
because we're doing effectRequests as a part of the selectionRequest (in
order to reuse the onClick logic in targettableEntity).

Vague todo list I had at the time of this commit:
add events for player hand requested  (so that we can get player hand
contents in TargetProvider without a reference to the component)
add events for cardSelectionRequestSupplied (Done, just need to get the
whole flow working through the state issues)
add casting cancellation check (can't have the player cancel casting
after seeing the deck contents/ordering or drawing off of the deck)

Biggest reason to pick this back up is consistency in casting
experience. Right now, any card targetting effect is uncancellable, and
the effects in procedures after them are. Basically, anything that
targets cards should be its own standalone procedure now.