Refactor entity architecture and enemy behaviors
This commit is large, apologies. This refactors the entity architecture
to make things component based instead of inheritence based. The biggest change
is that we now have CombatInstance, DeckInstance, CompanionInstance, EnemyInstance,
and MinionInstance. These are now all separate monobehaviors that have the logic
specific to them isolated in their respective components. Sometimes there's some overlap,
such as adding to the CombatInstance's OnDeath from the other components. It's done
in a modular way and I think it still follows good architectural patterns. Although
we need to revisit our triggers and event timing.

Additionally, this commit modified enemies to use the effect workflow system
that cards and compnion abilities use. This worked out pretty well.