Implement saga functionality, implement Con Artist
Added the ability to change the effect workflow a card will use on its
next cast.

Tested out this functionality + the cast count functionality on a whole
ONE saga so you know it has to work.

Feel free to try it out with the con artist card.

big TODO: Text box replacement on the cards based on which effect
workflow is being used, as a stand-in for the special saga card layout
we'll eventually have.