Add "View Companions" Button to the Shop
Duplicates the view companions button on the Map scene; all it does is call a method on the parent Manager to display the CompanionViewUI prefab.
This is temporary, throwaway work because we are planning on redoing the shop UI, but it is good practice for me to get used the Unity codebase.

Tests:
* shop shows up without breaking
* view companions button shows the current roster
* buy a companion, then view companions - shows the updated roster.