Fix duplicate companions Ability bug
I was seeing a problem where if you have multiple Bidoofs or Aborahs on your team, only one would get all the strength.
That is because the old implementation sets this.companionInstance on the Ability class - all the different Bidoofs are trying to write that companionInstance!
We should fix this so we do not have to maintain the "this.companionInstance" write reference on the Ability, but for now, a decent workaround is to create an "AbilityInstance" class that has a read-only reference to the Ability and keeps track of the companion.