Fix OutOfRange bug + deck selection for combining companions
There is a slight bug in the code, I got a count OutOfRange bug in the code off of main.
Fixed that.

I also made it automatically select the deck that is the largest out of the companions
being combined.
That means that you will get the one you added the most shop cards to, which should be
a decent proxy for what we are going for.

Tests:
combined my bidoofs into a gloriously big l2 bidoof