Assign cards based on CardPoolSO instead of CompanionTypeSO
When you have a Level1 and level2 companion of the same type, the
card purchasing in the shop gets weird.
The behavior we want is that the card pools for the both are the same,
and cards can be assigned to either level1 or level2.
It was especially annoying when you upgraded a companion and then tried to buy
a card that was in the shop - you couldn't until you refreshed.
This fixes that problem.

Test: tried it out