Tentative fix for effects not resolving before turn phases change
We have a few effect workflows that trigger when the turn phase changes, including
the Fireblob cards dealing damage at the end of combat.
Ethan noticed that the game would freak out a lil when the fireblob dealt fatal damage.
The combat would end, but the end of combat abilities would be cleared out by the
turn starting.
This should be an effective bandaid for now - I tested it out.
The current code does not work, because some of the coroutines we start do the naked
`StartCoroutine` instead of `yield return StartCoroutine`, so some of the effects
that we are trying to wait on are not part of the parent-child hierarchy.

Cheers,
James