Refactor and simplify Discard / exhaust effects
The code from before this commit is very confusing - there is no single entrypoint to exhaust or discard a card.
Now, the only entrypoint is through the PlayerHand.Instance.
Makes sense, ya?
This is also necessary to fix the Argos work because previously we did not wait until the coroutine completed.